update all query metrics
