+++ great_expectations/expectations/metrics/query_metrics/query_table.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
 
-from typing import TYPE_CHECKING, Any, ClassVar, Dict, List, Union
+from typing import TYPE_CHECKING, Any, Dict, List, Union
 
 from great_expectations.compatibility.sqlalchemy import (
     sqlalchemy as sa,
@@ -23,21 +23,6 @@ if TYPE_CHECKING:
 class QueryTable(QueryMetricProvider):
     metric_name = "query.table"
     value_keys = ("query",)
-
-    # The name of the parameter that will be used to pass the unexpected_rows_query to the query metric provider
-    _query_param: ClassVar[str] = "query"
-
-    @classmethod
-    def _get_query_from_metric_value_kwargs(cls, metric_value_kwargs: dict) -> str:
-        query: str | None = metric_value_kwargs.get(
-            cls._query_param
-        ) or cls.default_kwarg_values.get(cls._query_param)
-        if not query:
-            raise ValueError(
-                f"Must provide `{cls._query_param}` to `{cls.metric_name}` metric."
-            )
-
-        return query
 
     # <snippet>
     @metric_value(engine=SqlAlchemyExecutionEngine)
