+++ great_expectations/expectations/metrics/query_metrics/query_template_values.py
@@ -50,9 +50,7 @@ class QueryTemplateValues(QueryMetricProvider):
         metrics: Dict[str, Any],
         runtime_configuration: dict,
     ) -> List[dict]:
-        query = metric_value_kwargs.get("query") or cls.default_kwarg_values.get(
-            "query"
-        )
+        query = cls._get_query_from_metric_value_kwargs(metric_value_kwargs)
 
         selectable: Union[sa.sql.Selectable, str]
         selectable, _, _ = execution_engine.get_compute_domain(
@@ -110,9 +108,7 @@ class QueryTemplateValues(QueryMetricProvider):
         metrics: Dict[str, Any],
         runtime_configuration: dict,
     ) -> List[dict]:
-        query = metric_value_kwargs.get("query") or cls.default_kwarg_values.get(
-            "query"
-        )
+        query = cls._get_query_from_metric_value_kwargs(metric_value_kwargs)
 
         df: pyspark.DataFrame
         df, _, _ = execution_engine.get_compute_domain(
