+++ great_expectations/expectations/metrics/query_metrics/unexpected_rows_query_table.py
@@ -1,5 +1,4 @@
-from typing import ClassVar
-
+from great_expectations.compatibility.typing_extensions import override
 from great_expectations.expectations.metrics.query_metrics.query_table import QueryTable
 
 
@@ -7,5 +6,7 @@ class UnexpectedRowsQueryTable(QueryTable):
     metric_name = "unexpected_rows_query.table"
     value_keys = ("unexpected_rows_query",)
 
-    # The name of the parameter that will be used to pass the unexpected_rows_query to the query metric provider
-    _query_param: ClassVar[str] = "unexpected_rows_query"
+    @override
+    @classmethod
+    def _get_query_param_name(cls) -> str:
+        return "unexpected_rows_query"
